+++ src/pages/basket/api/apply-promo-code.tsx
@@ -4,8 +4,8 @@ import {
   saveTokenCookie,
   TOKEN_NAMES,
   type Cart,
+  type Actions,
 } from "../../../shared";
-import type { Actions } from "./types";
 
 // Передаём в функцию:
 //              добавить промокод:
