--- src/pages/basket/api/get-cart.tsx
@@ -1,25 +1,0 @@
-import { API_HOST, PROJECT_KEY } from "../../../project-config";
-import {
-  getTokenFromCookie,
-  saveTokenCookie,
-  TOKEN_NAMES,
-  type Cart,
-} from "../../../shared";
-
-export async function getCart(): Promise<Cart> {
-  let cart: Cart;
-  const BEARER_TOKEN = getTokenFromCookie(TOKEN_NAMES.successUserAccess);
-  const cartID = getTokenFromCookie(TOKEN_NAMES.cartID);
-  await fetch(`${API_HOST}/${PROJECT_KEY}/me/carts/${cartID}`, {
-    method: "GET",
-    headers: {
-      Authorization: `Bearer ${BEARER_TOKEN}`,
-    },
-  })
-    .then((response) => response.json())
-    .then((data: Cart) => {
-      saveTokenCookie(data.version.toString(), TOKEN_NAMES.cartVersion);
-      cart = data;
-    });
-  return cart;
-}
