+++ src/pages/basket/api/index.ts
@@ -1,4 +1,2 @@
-export { AddingDeletingModifyingItemsInCart } from "../../../shared/api/cart-api/adding-deleting-modifying-items-in-cart";
 export { deleteCart } from "./delete-cart";
-export { getCart } from "./get-cart";
 export { ApplyPromoCode } from "./apply-promo-code";
