+++ src/pages/basket/index.ts
@@ -1,2 +1,1 @@
-export { AddingDeletingModifyingItemsInCart, deleteCart } from "./api";
 export * from "./ui";
