+++ src/shared/api/cart-api/index.ts
@@ -1,3 +1,4 @@
 export { createCart } from "./create-cart";
 export { getCart } from "./get-cart";
-export { AddingDeletingModifyingItemsInCart } from "./adding-deleting-modifying-items-in-cart";
+export { addingDeletingModifyingItemsInCart } from "./adding-deleting-modifying-items-in-cart";
+export type { Actions } from "./types";
