+++ src/shared/api/index.ts
@@ -2,7 +2,8 @@ export { ObtainAnonymousAccessToken } from "./obtain-anonymous-access-token";
 export { sendingSignInOrSignUpRequest } from "./signin-or-signup-request";
 export {
   createCart,
-  AddingDeletingModifyingItemsInCart,
+  addingDeletingModifyingItemsInCart,
   getCart,
+  type Actions,
 } from "./cart-api";
 export * from "./types";
