--- src/shared/ui/cart/index.ts
@@ -1,1 +1,0 @@
-export { Cart } from "./cart";
