--- src/shared/ui/cart/styles.css
@@ -1,11 +1,0 @@
-.main-cart {
-  height: 50rem;
-  background-color: var(--background-color);
-}
-
-.cart-content {
-  font-size: 50px;
-  text-align: center;
-  color: var(--white-color);
-  padding-top: 100px;
-}
