+++ src/shared/ui/index.ts
@@ -1,6 +1,6 @@
 export { LoadingPlaceholder } from "./loading";
 export { Price, type PriceInfo } from "./price";
 export { NoResultsFound } from "./no-results-found";
-export { сreateProductCard } from "./product-card";
+export { createProductCard } from "./product-card";
 export { ButtonAddToCart } from "./button-add-to-cart";
 export { ToCatalogButton } from "./to-catalog-button";
