+++ src/shared/ui/product-card/create-product-card.tsx
@@ -6,7 +6,7 @@ import type { PriceInfo } from "../price";
 
 const FIRST_PRODUCT_IMAGE_NUMBER = 0;
 
-export function сreateProductCard(
+export function createProductCard(
   cardInfo: MasterData | Product,
 ): ReactElement {
   const {
